Extract `HistoryContainer`
