+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/ui/MedicalHistorySummary.kt
@@ -18,6 +18,7 @@ import org.simple.clinic.R
 import org.simple.clinic.common.ui.theme.SimpleTheme
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion
+import org.simple.clinic.medicalhistory.ui.HistoryContainer
 import org.simple.clinic.medicalhistory.ui.MedicalHistoryQuestionItem
 import org.simple.clinic.medicalhistory.ui.TobaccoQuestion
 
@@ -56,7 +57,7 @@ fun MedicalHistorySummary(
         strokeAnswer = strokeAnswer,
         kidneyAnswer = kidneyAnswer,
         diabetesAnswer = diabetesAnswer,
-        diabetesManagementEnabled = diabetesManagementEnabled,
+        showDiabetesQuestion = !diabetesManagementEnabled,
         onAnswerChange = onAnswerChange
     )
 
@@ -67,70 +68,6 @@ fun MedicalHistorySummary(
           showSmokelessTobaccoQuestion = showSmokelessTobaccoQuestion,
           onAnswerChange = onAnswerChange
       )
-    }
-  }
-}
-
-@Composable
-fun HistoryContainer(
-    heartAttackAnswer: Answer?,
-    strokeAnswer: Answer?,
-    kidneyAnswer: Answer?,
-    diabetesAnswer: Answer?,
-    diabetesManagementEnabled: Boolean,
-    modifier: Modifier = Modifier,
-    onAnswerChange: (MedicalHistoryQuestion, Answer) -> Unit,
-) {
-  Card(modifier = modifier) {
-    Column(
-        modifier = Modifier
-            .padding(horizontal = dimensionResource(R.dimen.spacing_16))
-            .padding(
-                top = dimensionResource(R.dimen.spacing_16),
-                bottom = dimensionResource(R.dimen.spacing_4)
-            )
-    ) {
-      Text(
-          text = stringResource(R.string.medicalhistorysummaryview_history),
-          style = SimpleTheme.typography.subtitle1Medium,
-          color = MaterialTheme.colors.onSurface,
-      )
-
-      Spacer(Modifier.requiredHeight(dimensionResource(R.dimen.spacing_4)))
-
-      MedicalHistoryQuestionItem(
-          question = MedicalHistoryQuestion.HasHadAHeartAttack,
-          selectedAnswer = heartAttackAnswer,
-          showDivider = true,
-      ) {
-        onAnswerChange(MedicalHistoryQuestion.HasHadAHeartAttack, it)
-      }
-
-      MedicalHistoryQuestionItem(
-          question = MedicalHistoryQuestion.HasHadAStroke,
-          selectedAnswer = strokeAnswer,
-          showDivider = true,
-      ) {
-        onAnswerChange(MedicalHistoryQuestion.HasHadAStroke, it)
-      }
-
-      MedicalHistoryQuestionItem(
-          question = MedicalHistoryQuestion.HasHadAKidneyDisease,
-          selectedAnswer = kidneyAnswer,
-          showDivider = !diabetesManagementEnabled,
-      ) {
-        onAnswerChange(MedicalHistoryQuestion.HasHadAKidneyDisease, it)
-      }
-
-      if (!diabetesManagementEnabled) {
-        MedicalHistoryQuestionItem(
-            question = MedicalHistoryQuestion.DiagnosedWithDiabetes,
-            selectedAnswer = diabetesAnswer,
-            showDivider = false,
-        ) {
-          onAnswerChange(MedicalHistoryQuestion.DiagnosedWithDiabetes, it)
-        }
-      }
     }
   }
 }
